requirements.txt
asteval==1.0.6
astropy==7.0.1
astropy-iers-data==0.2025.2.17.0.34.13
astropy_healpix==1.1.2
casaconfig==1.3.1
casatasks==6.7.0.31
casatools==6.7.0.31
certifi==2025.1.31
cftime==1.6.4.post1
click==8.2.1
cloudpickle==3.1.1
colour==0.1.5
colour-science==0.4.7
contourpy==1.3.1
corner==2.2.3
crc32c==2.7.1
cycler==0.12.1
dask==2025.5.1
dill==0.4.0
donfig==0.8.1.post1
emcee==3.1.6
fonttools==4.56.0
fsspec==2025.5.1
imageio==2.37.0
importlib_resources==6.5.2
joblib==1.5.2
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.44.0
lmfit==1.3.3
locket==1.0.0
matplotlib==3.10.0
netCDF4==1.7.2
networkx==3.5
numba==0.61.2
numcodecs==0.16.1
numpy==2.2.3
packaging==24.2
pandas==2.2.3
partd==1.4.2
pillow==11.1.0
pyerfa==2.0.1.5
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
reproject==0.14.1
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.2
six==1.17.0
threadpoolctl==3.6.0
tifffile==2025.9.20
toolz==1.0.0
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.1
uncertainties==3.2.3
zarr==3.0.8
